backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "alembic>=1.18.3",
    "asyncpg>=0.30.0",
    "fastapi>=0.128.0",
    "fastapi-users[sqlalchemy]>=15.0.3",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.40.0",
]
